backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
python-i18n==0.3.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2 